backend/requirements.txt

Django
djangorestframework
django-cors-headers
Pillow
python-decouple